Finalizando Busca 1

+++ Trabalho Fase 2/Hash.h
@@ -14,7 +14,7 @@ class Hash
     
 
     private:
-        int tamanho;
+        long long int tamanho;
         ListaEncad** registros;
 };
